DL-13367: Refactoring tests
